Commit to sync with new desktop

+++ core/src/main/java/ore/forge/Direction.java
@@ -1,18 +1,18 @@
 package ore.forge;
 
 public enum Direction {
-    EAST(0),//x+1, y+0
+    EAST(0.0f),//x+1, y+0
     NORTH(90),//x+0, y+1
     WEST(180),//x-1, y+0
     SOUTH(270);//x+0, y-1
 
-    public final int angle;
+    public final float angle;
 
-    Direction(int angle) {
+    Direction(float angle) {
         this.angle = angle;
     }
 
-    public int getAngle() {
+    public float getAngle() {
         return angle;
     }
